refactor(system-tests): move the dev root CA into its own module
A pure move, no behaviour change. The dev root CA is about to gain a second
consumer -- the local backend, which installs it into the driver's own trust
store -- and it belongs to neither that nor the gateway: what trusts it is the
dev IC-OS images. Putting it in `driver/dev_root_ca.rs` keeps the certificate
issuance machinery (`rcgen`, `rsa`) out of the QEMU/networking module, which
would otherwise own code it never calls.
